Add animations and bug fix sleeping
Animations are things that appear twice in the game. One: Monsters will be able to breath cones of fire (or other  types, but none of this is implemented) which will only be a dragon for the base 26 monsters. The other way, is that 3 wands have area attacks, and they make explosions in the color of the type of wand.