Fix game-breaking issues with converting to JAR file
Now the game can be distributed to beta testers. Many bugs are still unresolved from the switch to JAR file, and those will need to be addressed.

Noticed Bugs:
There is a strange white bar that appears on screen and only dissapears when you click the window.
The game cannot be closed and must be quit forcefully through Task Manager

Additional Note:
The program is not functional unless the "data" directory (sub-directory to the resources root directory) is copied into the same folder as the JAR file. This is a problem for another day